requirements.txt
jupyter
pytest
pytest-pycodestyle
request